requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
pandas==1.5.3
numpy==1.24.3
matplotlib==3.7.5
seaborn==0.13.2
pymysql==1.1.0
python-dotenv==1.0.1
Werkzeug==3.0.1
openpyxl==3.1.2
plotly==5.18.0 